refactor: 상품 컴포넌트 자판기 페이지에서 분리
