+++ src/Component/index.js
@@ -2,3 +2,4 @@ export {InsertButton} from './InsertButton';
 export {Products} from './Products';
 export {UiChangeBtn} from './UiChange';
 export {VendingMachineInterface} from './VendingMachineInterface';
+export {ProductList} from './ProductList';
